chore: resolve unexpected ring dependencies
